requirements.txt
fastapi>=0.103.0
python-multipart>=0.0.6
uvicorn>=0.23.0
onnxruntime>=1.15.0
pydantic>=2.3.0
pytest>=7.4.0
httpx>=0.25.0
Pillow>=10.0.0
